requirements.txt
mapillary
python-dotenv
requests
pandas
geopandas
h3
pyarrow
pillow
duckdb